+++ substrabac/substrapp/signals/objective/post_delete.py
@@ -8,4 +8,5 @@ def objective_post_delete(sender, instance, **kwargs):
     instance.metrics.delete(False)
 
     directory = path.join(getattr(settings, 'MEDIA_ROOT'), 'objectives/{0}'.format(instance.pk))
-    shutil.rmtree(directory)
+    if path.exists(directory):
+        shutil.rmtree(directory)
